+++ HISTORY.rst
@@ -10,7 +10,7 @@ Added routines to compute coordinates for the graphs.
 
 Added fast filtering of ideal band-pass.
 
-Implemented spectrogramm for graphs.
+Implemented graph spectrogramms.
 
 Added the Barabási-Albert model for graphs.